refactored draw_line
